🐛 Ensure state observers are notified of errors before the machine resets
The state machine removes observers when the state is set to idling, so a transition to .idling with a side effect of an error meant the observer that cares about the error in the side effect was removed just before the side effect executed.

Now rather than transition to .idling, transition with no state change and let the side effect change the state when necessary.