♻️ Refactor ExperienceRenderer to support multiple ExperienceStateMachine instances for non-modal Experiences
There is one sharedStateMachine for modal experiences and a mapping by experience ID to state machine for non-modal.

♻️ Add the experience ID to the config mapping sent to ExperienceActions during execution

This update is to support actions being able to pass the experience context to the ExperienceRenderer, and operate on the correct state machine for that Experience. It uses the instance ID of the experience in the chance that the same experience could be rendering in multiple places.

✨ Add the AppcuesEmbedTrait and support for embedded content in AppcuesView

🎬 Update example app to show embedded experiences

👌 Updates for content embed from PR review

👌 Restore the order of trait application, moving presenting trait to last

👌 Fix AppcuesLaunchExperienceAction action